Send some empty packets before reset request
If the target is sending some kind of repetitive data, then it's likely
that the packet pool will be empty when we start up. Make sure we send a
few empty packets first, to free up some space in the packet pool (the
packets waiting to be transmitted on the target will go back to the free
list).